requirements.txt
psycopg[binary]==3.2.3
pandas==2.2.3
python-dotenv==1.0.1
